chore: add senderStatusCard and ReceiverStatusCard to StreamDetails for PR #30
